+++ packages/desktop/src/renderer/components/SearchModal/ChannelItem.tsx
@@ -1,0 +1,41 @@
+import React, { useEffect, useState } from 'react'
+import { Typography } from '@mui/material'
+import classNames from 'classnames'
+import { useEnterPress } from '../../containers/hooks'
+
+const ChannelItem = ({
+  item,
+  focused,
+  className,
+  classNameSelected,
+  onClickHandler,
+  channelInput
+}) => {
+  const [initialRender, setInitialRender] = useState(false)
+
+  useEffect(() => {
+    setInitialRender(true)
+  }, [])
+
+  useEnterPress(() => {
+    if (focused) {
+      onClickHandler(item.address)
+    }
+  }, [focused, channelInput])
+
+  return (
+    <div
+      key={item.name}
+      className={classNames(className, {
+        [classNameSelected]: focused
+      })}
+      tabIndex={0}
+      onClick={() => {
+        onClickHandler(item.address)
+      }}>
+      <Typography variant='body2'>{`# ${item.name}`}</Typography>
+    </div>
+  )
+}
+
+export default ChannelItem