+++ packages/desktop/src/renderer/components/ui/ErrorModal/ErrorModal.test.tsx
@@ -23,7 +23,7 @@ describe('ErrorModal', () => {
           aria-hidden="true"
         />
         <div
-          class="Modalroot MuiModal-root css-1hk9oxe-MuiModal-root"
+          class="Modalroot MuiModal-root css-1voaj9u-MuiModal-root"
           role="presentation"
         >
           <div