+++ packages/desktop/src/renderer/sagas/modals/modals.types.ts
@@ -16,5 +16,7 @@ export enum ModalName {
   createCommunityModal = 'createCommunityModal',
   sentryWarningModal = 'sentryWarningModal',
   loadingPanel = 'loadingPanel',
-  uploadedFileModal = 'uploadedFileModal'
+  uploadedFileModal = 'uploadedFileModal',
+  leaveCommunity = 'leaveCommunityModal',
+  searchChannelModal = 'searchChannelModal'
 }