+++ packages/desktop/src/renderer/theme.ts
@@ -91,6 +91,7 @@ export default createTheme({
       greenDark: '#9BD174',
       red: '#FF0000',
       hotRed: '#E42656',
+      hotPink: '#E42656',
       linkBlue: '#59c0d5',
       buttonGray: '#E3E3E3',
       black30: '#4C4C4C',