+++ packages/mobile/android/app/src/main/java/com/zbaymobile/Scheme/WebsocketConnectionPayload.kt
@@ -1,5 +1,0 @@
-package com.zbaymobile.Scheme
-
-data class WebsocketConnectionPayload (
-    val dataPort: Int
-)